Podcast data, some HTML cleanup
The latest podcast episode name, description, and link are now pulled
from FeedBurner. The background image is still static because there
isn’t a standard system for podcast images yet.

Instead of using a PHP compiler (which in this scenario is difficult to
work with/test for multiple reasons: for example github.io won’t
execute the PHP, and my localhost doesn’t seem to like it either) the
proxy work is offloaded to Yahoo Query Language.

Miscellaneous HTML cleanup.